Add test coverage for async slicing of points and fix rounding bug (#5783)
# Description

This PR adds some async slicing integration tests for points, one of
which was failing due to unintended rounding. It also adds some debug
logs across sync and async slicing paths to make it easier to debug
issues. Finally, it changes the main log message format to include the
thread name, which is incredibly helpful when debugging concurrency
issues.

# References

This was split off from #5711 to avoid bloating that already large PR.
Credit to @kcpevey, who found the the bug and wrote most of #5711. Also
this part of the larger #4795

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# How has this been tested?
- [x] added new tests to cover issues
- [x] all existing tests pass with my change

I manually tested behavior and logging with async off and on.