Do not blend color in QtColorBox with black using opacity (UI suggestion; bugfix? ) (#5827)
# Description
For some reason that I cannot fully understand, the displayed color in
QtColorBox is multiplied with the layer opacity, which is equal to
blending the selected color with black. I find it very misleading as it
just distorts displayed colors making them harder to associate with the
true colors. It becomes especially confusing when you use low levels of
opacity (20-30%), which makes all colors look grey and
indistinguishable.

How it now looks like in napari depending on different opacity:

![image](https://github.com/napari/napari/assets/5459767/01b55cfd-6ce4-494d-a0fa-a8116ee35d6b)

This PR just removes it making it always show the actual color.

P.S. It has been puzzling me for a while why colors are always darker
than the real colors in the color box. I've only recently realized that
their appearance is depending on the opacity.

## Type of change
- [X] Bug-fix (non-breaking change which fixes an issue)

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality