Add test to check basic interactions with layer controls widgets (#5757)
# Fixes/Closes
Initial effort towards improving GUI testing automation:
https://napari.zulipchat.com/#narrow/stream/212875-general/topic/GUI.20testing

# Description
This is an initial attempt to add more tests for code related with GUI
elements. The new parametrized test proposed creates an instance of a
layer (`Image`, `Labels`, `Points`, `Shapes`, `Surface`, `Tracks` or
`Vectors`) and then operates over the layer control widgets simulating
user usage (`QCombox`, `QSpinBox` and `QSlider` based widgets value
changes, clicking `QCheckbox` and buttons widgets, `QColorSwatch`
widgets color changes typing colors, etc)

This also fixes a little issue that I noticed while working on the test.
The error is related with the value of the color mode combobox for a
Vectors layer controls. The combobox was not showing the correct value
after a fallback to a valid value is done (when a `ValueError` exception
is raised, the color mode value of the layer is changed to the previous
one, but the combobox value was missing being updated too):

* Before:

![vectors_before](https://user-images.githubusercontent.com/16781833/234071199-5d22d1b4-0653-489a-be92-68ca138033a0.gif)

* After:

![vectors_after](https://user-images.githubusercontent.com/16781833/234070708-52a41085-eea8-4cd9-bcde-03c5b71395b5.gif)

Besides that, the test structure also has some logic to skip some
tracebacks from issues which are still open/were discovered while
writing the test like #5798 and #4864

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Add more tests

# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [x] all tests pass with my change
- [x] I check if my changes works with both PySide and PyQt backends
      as there are small differences between the two Qt bindings.  

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas