MAINT: Try to bump Python versions in CI.
It seem that it's been some time since Python 3.11 is out, Try to use it
where relevant, and extend out test suite to test on Python 3.11. Where
relevant also use "3.x" which means latest stable, so it should
auto-bump to 3.12 when released, which avoids an extra bump.

I'm hopping this will slightly speed-up CI as Python 3.11 is reported to
be quite faster in some use case, but at least let us know if there are
things we are waiting for for for 3.11 compatibility.