Fix usage of closed interval dims and unify extents calculations (#5751)
# Description
See [this zulip
thread](https://napari.zulipchat.com/#narrow/stream/212875-general/topic/Dims.20in.20world.20space)
for some background on the issue. In short, after the change from
"slider space dims" to "world space dims", some things required some
changes that we didn't address in #5522. This PR tries to solve them.

*TLDR; all layers now have `extent` and `_extent_augmented` (reflecting
the existing partial implementation in `Points`) to distinguish between
the extent of the actual data (coordinates for points, pixel samples for
images) and the extent including also the "size" of the visualisation of
that data point (i.e: size of points and pixels).*

Image extents in data space were actually always +1 than they "should"
be; if we consider pixels to be data points [rather than little
squares](http://alvyray.com/Memos/CG/Microsoft/6_pixel.pdf) (as we
should) and whose origin is at `(0, 0)`, an image of shape (100, 100)
extends from 0 to 99, and not from 0 to 100 (as is currently in `main`).

This then translates in the `world_extent` being off by `1 * scale`,
which until now wasn't a problem because `Dims` was using integers as
the source for `point` and we considered `dims.range.stop` to be
exclusive (it is now *inclusive* since #5522!). This is no longer true
becase the source for `point` is floats in world coordinates, so you can
get arbitrarily close (and also equal) to the end.

On top of this, we were shifting the image extents by 0.5, only for then
re-shifting it when we actually needed to calculate the dims extent.
This is unnecessary if we simply never treat pixels as little squares.

A consequence of these changes is also that `Layer.corner_pixels` (used
mostly in multiscale and octree) are affected similary: an image of
shape `(1, 2)` used to have `[1, 2]` as bottom right pixel, but now it's
`[0, 1]`.

```
import napari
import numpy as np
img = np.random.rand(5, 4, 6)
pts = np.array([[0, 0, 0], [5, 6, 3]])
v = napari.Viewer(ndisplay=3)
il = v.add_image(img, translate=[-1, -1, -1])
pl = v.add_points(pts, size=1, scale=[0.3, 0.5, 1])
v.dims.range
```

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>