Add secondary keybindings to restore mnemonic (non-numeric) shortcuts in layers (#5155)
In #3072, we updated shortcuts in the labels layer to be more ergonomic,
as described in #2800. However, we lost the ability to easily remember
which shortcut does what, by assigning them to numbers instead of
letters associated with the action. In this PR, we add back those
shortcuts as secondary options, so that users can use both the adjacent
"ergonomic" controls and the more mnemonic shortcuts associated with
letters.

Supersedes #4696

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>