[Qt6] Fix AttributeError on renaming layer (#5850)
# Fixes/Closes

Closes #5849

# Description

This is a pretty simple fix, but is also backward-compatible in my
testing with Qt5. I added a test to cover this method for the future. I
can look into adding more tests for that file as well.

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# How has this been tested?
Added a test for this feature that fails without
a062d7933ccb6c0aeead275a3418b5c478615e30.

This test passes with a062d7933ccb6c0aeead275a3418b5c478615e30 on both
Qt5 and Qt6.

## Notes

@psobolewskiPhD:

> This works for me in both pyqt6 and pyqt5.
> The fix looks correct to me, which can be confirmed reading:
> doc.qt.io/qtforpython-6/considerations.html#forgiveness-mode-and-type-hints