Add instructions to build napari docs on Windows (#158)
# Description
Added instructions on how to build the napari docs on Windows using
either [Git Bash](https://gitforwindows.org/) or
[WSL(2)](https://learn.microsoft.com/en-us/windows/wsl/install).

## Type of change
- [x] Fixes or improves existing content

# References
closes #19

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality

I think so? Another option is to make a cross platform makefile (see
[here](https://github.com/seankmartin/napari-docs/blob/cross_platform_make/Makefile))
but that seemed like it would introduce a lot of complexity to the
Makefile. Git bash and WSL both work very well for building the docs and
are probably good tools to get used to anyway.

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Melissa Weber Mendonça <melissawm@gmail.com>