Bugfix: link the Labels model to the "show selected" checkbox (#5867)
# Description
This PR adds the missing connection between the "show selected" checkbox
in the Qt Labels controls and the Labels layer. If you modify the state
of `labels_layer.show_selected_label` in code without it, the checkbox
is not synced with the true state. For example, the similar checkbox
"preserve labels" is connected to the model, so I just added it for
"show_selected_label" in the very same fashion.

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>