Add lasso tool for faster drawing of polygonal Shapes (#5555)
# Description

This PR adds a polygon lasso tool for drawing complex polygons which are
added to a shape layer. The tool does not require the user to click for
adding each specific vertex as is the case with the polygon tool. This
PR features the following:
• A new shape layer button which can be activated by keyboard shortcut
shift + P
• A polygon lasso mode which can be used for drawing with the mouse and
tablet

For mouse draw mode the sequence of events by the user is as follows:

1. Mouse press
2. Move mouse while not mouse pressed
3. Mouse press to finish draw

For tablet mode, the shape is created by a mouse (ie stylus) click+drag
event. Drawing finishes when the stylus is lifted / on mouse release.

For both modes, vertices are added only if the vertex to be added is at
least 10 screen pixels away from the previous vertex. This is a setting
configurable in advanced settings.

An implementation of the Ramer–Douglas–Peucker algorithm is used to
reduce the number of vertices when a drawing is finished. The
aggressiveness of reducing the number of vertices is dependent on a
parameter epsilon. If set to 0, no vertices are removed. The higher the
epsilon, the more vertices are removed. The epsilon parameter can be
configured in napari preferences -> experimental -> rdp_epsilon.

Lastly, a refactor has been performed for various functions related to
polygon drawing.

# References

We acknowledge @yozhikoff who made a first prototype of this new
feature.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: wmv_hpomen <w-mv@hotmail.com>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>