Add a napari plugin debugging quick start section to the debugging guide (#161)
# Description
Adds a "quick start" section to the existing napari plugin debugging
guide that demonstrates writing a launch script with your plugin setup
so that you can debug it.
I've referenced VSCode here quite a bit because that the IDE I'm most
familiar with, it has decent debugging documentation, it is free, and
I'm not sure how much I'd recommend using pdb straight off the bat!

## Type of change
- [x] Fixes or improves existing content

# References
closes #154

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have added [alt text](https://webaim.org/techniques/alttext/) to
new images included in this PR

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>