Mypy error checking opt-in to opt-out.
Currently mypy configuration is opt-in, where we ignore errors by
default.

This it seem like we should do the opposite so that when new code / new
file are added they are checked, and them progressively work our way
down not adding typing errors.

If we want global ignore of some folder we can add it but be explicit.