Fix drawing vertical or horizontal line segments in Shapes layer (#5895)
Closes #4538.

In #4538, @psobolewskiPhD reported that *something* prevented drawing of
Shapes line segments that were either perfectly vertical or perfectly
horizontal. This was caused by faulty logic meant to prevent drawing
segments that are *too small*. This changes the logic to prevent
changing a scale to 0, which fixes the bug. It also renames some
variables to make the logic easier to follow.