MAINT: Start testing on Python 3.11 in CI. (#5439)
Python 3.11 was released in October 2022. 
In this PR, the CI test suite is extended to test on Python 3.11. 

I'm hoping that CI on Python 3.11 will be faster, as 3.11 is reported to
be quite a bit faster in some use cases.

Additionally, there are a few refactors and split of test into multiple
subtests, as some of them are segfaulting on some platforms/python
version combinations.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>