TEST: Fixture to "uncreate" IPython.
As pointed out by #5886 we should not rely on test ordering and test
that create and IPython instance should try to destroy it.

This is my attempt at doing this.

It's not perfect, and need explict use but can be the start of a proper
fixture.

We could have an autouse fixture that make sure
InteractiveShell._instance is None at the end of every test to make sure
all is properly cleaned. I'm not sure it's worth it.