MAINT: GitHub action Timeout + 5 minute
The timeout is currently 35 minutes. The macos-latest Python-3.9 pyqt5
run-time is one of the slower one, close to the limit. Some of the
recent PRs where this item is passing run in almost 35 minutes. E.g PR 5926 in 33m 26s

It's not unexpected to have variation of the time that would put us
above the threshold of 35 minutes.

The timeout was added as some builds were stalling for 6 hours, but
failing at the 35 minute mark does not seems like representative of
stalling.

Thus bumping to 40 minutes.

Ideally it would be good to have a larger timeout only on macos
platform, but I don't believe timeout-minutes supports expression.