Enable pinch-to-zoom with vispy 0.13+ (#5859)
# Description
This enables pinch-to-zoom in napari when using a trackpad, so long as
you have vispy 0.13+ installed. As this functionality is now built-in
with the main vispy cameras, the behavior sort of comes to napari for
free, but this change is required because we modify the vispy camera
classes. See
https://github.com/napari/napari/issues/166#issuecomment-1552743876 for
some additional discussion.

I think part of the reason this feature could be useful is the potential
to bind scrolling (`mouse_wheel`) and pinching (`gesture_zoom`) to
different behavior. For example, I recall @psobolewskiPhD mentioned (on
Zulip?) wanting to use pinching to zoom and scrolling to change brush
size. I believe it will requires additional work to enable such
capabilities.

# References
https://github.com/vispy/vispy/pull/2456
https://github.com/napari/napari/pull/5701