Improve documentation of `changed` event in EventedList (#5928)
(Partially addresses #5927)

# Description

The docstring of `EventedList.events.changed` is not very clear
and has led to confusion. This PR updates it to make it clear that
the event is only triggered when an item value at a specific index
is changed, rather than when the list itself is changed (for
example, with insertion or deletion of elements).

A typo is also fixed.