Push variables to console when instantiated (#5210)
Currently, if
[`update_console`](https://github.com/napari/napari/blob/7117d22e4baa740864762454911dc5ed09764ccc/napari/viewer.py#L88)
is called on viewer object before the console is instantiated the
function does nothing. With this PR instead references to the requested
variables are added to a `console_backlog` list attribute on qt_viewer.
If and when console is instantiated these variables are pushed.

Unsure if [this related
PR](https://github.com/napari/napari-console/pull/18) is moving forward
and whether it would solve this case.

- [x] Enhancment

Closes #4098

- [x] Ran unit tests for `napari/_qt`. There were some errors but looked
unrelated to my changes.
- [x] Added unit test based on existing `test_update_console`

- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>