Skip equality comparisons in EventedModel when unnecessary (#5615)
# Description
Alternative to #5614 as I cannot fit in the form of review with this. 

Evented model checks if a field value (and since
https://github.com/napari/napari/pull/5528 also all the field's
dependent properties) has changed before firing an event. This
comparison may be expensive in some cases.

In this PR, the comparison is not performed if there is no connected
callback to the event.

Also the eq operator for properties with type annotation is determined
in `EventedMetaclass.__new__`

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] New feature (non-breaking change which adds functionality)


Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>