Disallow outside screen geometry napari window position (#5915)
# Fixes/Closes
Closes #5912 

# Description
If napari is closed while windows (not full screen) are on an external
display, then upon reopening when the external screen is not connected,
the napari window will not be accessible. This is due to the saved
window position in the settings (from the external screen) being
off-screen (the x or y position is negative) with just one screen. This
PR fixes this by checking how many screens are available and if only 1,
ensuring the position of the window is set to default values if saved
position is negative or falling outside the screen geometry.

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# How has this been tested?
Added test for the `_load_window_settings` method of the `_QtMainWindow`
class. Upon opening napari, this method is used to get a valid window
position for napari.

Also tested manually. For manual I checked both with the external screen
set to be to the left and above the primary screen. For additional check
I also checked with secondary screen to the right and below the primary
screen although these cases were already covered by the code.

For each condition the approach is:
1. Get napari in windowed mode on the external display and close.
2. Disconnect external screen
3. Reopen napari and see if it is properly displayed on the main screen

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>