Emit extra information with layer.events.data (#5967)
# Fixes/Closes
Closes #5962, #2809, 

# Description
This PR lets extra information be emitted with `layer.events.data`. In
particular, it emits `action` which is either `add`, `remove`, or
`change` indicating what the action is on a particular shape or point.
Here, `change` is a union of move / drag events and vertex insert and
remove events. Next to `action`, `data_indices` is emitted indicating
which indices of `layer.data` are affected. This is of the form
`tuple[int]`. Lastly `vertex_indices` can be emitted in case of a
`change` action on a shape. This is of the form `tuple[tuple[int]]`. This
is to account for moving multiple shapes resulting in all vertices of
multiple shapes being changed. Using a `tuple` here ensures that order is
maintained in both `data_indices` and `vertex_indices` so their order is
the same and they can be enumerated over at the same time.

In short, this makes the emitted information a lot more useful for
plugin developers who might want to make adjustments to external data
based on what is happening to the data in a particular layer.