Replace old async loading with new approach (#5816)
# Fixes/Closes

Closes #5517
Closes #2790

# Description

This replaces [the old approach to asynchronous
loading](https://napari.org/0.4.16/guides/rendering.html) with [the new
approach described in
NAP-4](https://napari.org/stable/naps/4-async-slicing.html).

See the old approach (on `main`) on some remote 3D multi-resolution
images from Janelia via [@chili-chiu's fantastic bio sample data
plugin](https://www.napari-hub.org/plugins/napari-bio-sample-data):

https://github.com/napari/napari/assets/2608297/e8673d4e-c387-49bc-9254-30d0fc9c1cd2

And the new approach on the same data that is more responsive:

https://github.com/napari/napari/assets/2608297/0b9f1f12-d5e0-4980-be05-c5c58285c1a2

This PR also removes the old approach to tiled rendering (AKA octree)
since that is tied to the old approach to asynchronous loading. Most of
the code change delta consists of deleted files that are experimental
modules and tests associated with those old approaches.

The other changes remove the old integration and adjust the new one to
effectively achieve replacement. The key changes are as follows.

1. Remove async/sync_only marks from pytest since these should only be
needed for the old approach.
2. Remove command that hides/shows whether octree chunks are visible.
I'm unsure if these commands are publicly exposed anywhere, though this
one currently appears to be broken.
3. Redefine the existing experimental setting previously used to enable
the old approach to enable the new one instead.
4. Add support for `Layer.loaded` for the new approach by adding an ID
to each slice request. This could probably be moved to a separate PR if
desired (e.g. see https://github.com/andy-sweet/napari/pull/51)
5. Absorb slicing logic spread across old classes into the slice request
call.

This PR currently includes #5783, which should be reviewed and merged
first to make this large PR as small as possible. For the difference
between this and #5783, see https://github.com/andy-sweet/napari/pull/50

# References

I think this improves the simplicity of slicing logic as described in
#2156
Also see #5711 for the first version of this PR from @kcpevey, which was
split into this and #5783
Part of #4795

## Type of change

There probably need to be some documentation updates related to this
since the old approach is well documented on napari.org. We should
probably write some developer docs based off the async slicing NAP and
mark the old docs as deprecated and/or remove them to avoid confusion.

# How has this been tested?

- [x] added an integration test for checking `Layer.loaded` is updated
correctly
- [x] all existing tests pass with my change

I also did some manual testing by running `NAPARI_ASYNC=1 napari` and
interacting with some remote and non-remote data.

---------

Co-authored-by: Kim Pevey <kcpevey@gmail.com>
Co-authored-by: Kim Pevey <kcpevey@quansight.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Matthias Bussonnier <bussonniermatthias@gmail.com>