Keep shading light for Surface fixed relative to camera (#5893)
# Description

Currently, if `Surface` shading is set to `flat` or `smooth`, the light
is fixed in *data space*. This makes it hard to see any features on the
"shadow" side of the surface.

This PR updates the light position every time the camera moves, in order
to always keep it "behind the camera on the top right". We could
potentially expose this at some point, but this is a net positive
compared to what we have now.

Here's the difference between this PR (above) and main (below):

https://github.com/napari/napari/assets/23482191/9be5ebe4-7751-40f7-900c-c5765bf2c7bd

https://github.com/napari/napari/assets/23482191/39ce8f22-2694-4488-b26c-cbf097fc0526

I'm not 100% on the implementation, because I had to add a connection
between the camera and the layers, which is normally not there, so it's
entangling things a bit more... But I'm not sure there's a smarter way
to deal with it. One good thing is that the layer *model* is not
involved at all, so this is not increasing the "layer state" burden.

# References

- #2407