fix various keybinding/shortcut bugs (#5604)
This PR addresses bugs where the improper symbols would be displayed for
e.g. ⌘V on Macs, as well as improper capturing on Macs where e.g.
Control would be read as Meta and vice versa. This relies on an upstream
fix in `app-model` and uses the logic there to properly capture the
correct modifier keys, regardless of if Qt is switching Control with
Meta or not.

fixes #5047 
fixes #5332 
fixes #5579 

depends on https://github.com/pyapp-kit/app-model/pull/82