Fix behavior of PublicOnlyProxy in setattr, wrapped methods, and calling (#5997)
# Fixes/Closes

Closes #5767

# Description

This PR makes a few tweaks to the behavior of PublicOnlyProxy:

- in setattr, if we want to set an attribute of a wrapped object, if the
attribute value is itself an object wrapped with PublicOnlyProxy, we
unwrap it. This has two benefits: (1) checking the attribute later
incurs a significant cost in `_is_called_from_napari`, which involves
frame inspection. This way we don't incur that cost. And (2) Certain
equality checks fail when PublicOnlyProxy's interact with Qt. This is
#5767 and this PR fixes that.
- in `PublicOnlyProxy.create`, we now check whether the object we want
to wrap is a method, because then the module we need to check is not
that of the method (this is always `builtins`), but rather the module of
the object *that that method is bound to.* This is now fixed.
- When a PublicOnlyProxy'd object is callable, we unwrap the arguments,
call the unwrapped function with the unwrapped arguments, then wrap the
result. This avoids the performance issues detailed above.

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>