Add arrow display styles to Vectors layer (#5740)
# Fixes/Closes

Closes #4149

# Description

This is a follow-up of the discussion #4149 in which adding different
display style to the vectors of the Vectors layer was requested. As
suggested by users @brisvag and @jni, I implemented a ComboBox in the
GUI with 3 choices:

1. `rectangle` which is the old style of vectors, and currently the
default value
2. `triangle` which displays vectors as triangles
3. `arrow` which displays a "classical" vector shape, with a rectangular
shaft and a triangular head

![napari_vectors_display_styles](https://user-images.githubusercontent.com/94963445/232590807-9f812cfc-293a-4bd5-87ac-8b7acfdc7b87.gif)

Most of this contribution comes from additional functions to compute
vectors meshes in _vispy/layers/vectors.py. Because changes are mainly
in the function `generate_vector_meshes_2D`, there was no need to change
anything in the functions `_on_data_change` or `generate_vector_meshes`,
i.e it works natively with the 3D visualization.

The Vectors layer now provides a parameter `edge_display_style`.
Right now the possible values are: `[rectangle, triangle, arrow]`

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>